+++ nc1020.h
@@ -14,7 +14,7 @@ typedef struct WqxRom WqxRom;
 extern void Initialize(WqxRom);
 extern void Reset();
 extern void SetKey(uint8_t, bool);
-extern void RunTimeSlice(size_t, bool);
+extern void RunTimeSlice(uint32_t, bool);
 extern bool CopyLcdBuffer(uint8_t*);
 extern void LoadNC1020();
 extern void SaveNC1020();
